refactored Interfaces (added them inside the corresponding pages or redux slices
