+++ client/src/pages/Register/Interfaces/RegisterDataInterface/RegisterDataInterface.ts
