+++ client/src/store/slices/AuthSlice/Interfaces/UserInterface/UserInterface.ts
