+++ client/src/store/slices/AuthSlice/Interfaces/UserInterface/index.ts
