+++ client/src/store/slices/AuthSlice/index.ts
@@ -1,0 +1,1 @@
+export * from './AuthSlice';
